add initial lm32 test bench
Add a first draft of a test bench for the processor. The design can be
simulated with Icarus Verilog.

There are three plusargs "+prog=vhfile", "+dump=vcdfile" and
"+trace=tracefile". The first is mandatory and specifies the program which
is loaded into memory, while the second and the third options are optional.
When given, a VCD dump rsp. a trace file will be generated. The program
file can be generated with "objcopy -O verilog".

The memory has one write and two read ports. With this scheme, the majority
of the use cases are covered. Just the JTAG debug won't work because it
uses the write port of the instruction bus, which is not connected in the
test bench.

Signed-off-by: Michael Walle <michael@walle.cc>