add tlb control registers
Add TLBVADDR, TLBPADDR and TLBBADVADDR registers. These are used to
control the data and instruction TLBs.

Control functions like flushing and invalidating an entry are executed by
writing to the TLBVADDR register. An entry can be updated by writing to the
TLBPADDR (before writing the virtual frame number to TLBVADDR, with the
lower bits zero to indicate a NOOP).

The lowest bit in the TLBVADDR and TLBPADDR register is used to address the
corresponding TLB.

After a TLB exception has happened, the TLBVADDR register is automatically
preloaded with the missed virtual frame number. Because of this, an
exception handler can read the missed virtual address by reading
TLBBADVADDR register and set the physical frame number by writing the
TLBPADDR register.

Signed-off-by: Michael Walle <michael@walle.cc>
Signed-off-by: Yann Sionneau <yann.sionneau@gmail.com>