add PSW register
Add the processor status word register. This register is used to enable the
TLBs and the (non-privileged) user mode. If an exception occurs, the these
enable bits will be saved and cleared. After an eret or bret opcode is
executed, these bits will be restored.

For backward compatibility, the lower three bits of this register are
shadowed from the original IE register. A new program can then only use
this register, eg. saving the state on an ISR, while older programs can
still use the IE register.

Signed-off-by: Michael Walle <michael@walle.cc>